Always open log files as cp1252
Even if the system-default text encoding is something else, apparently script extender logs always use cp1252. Characters unsupported by cp1252 end up converted into their closest equivalent (or, in some cases, the same character with the accent removed - `ÿ` is `0xFF` in cp1252, but load a DLL with that in the name, and you'll get `y` in the log).

I'm still not 100% convinced that this is the right encoding (I'd really like the log from this guy that actually discovered the issue https://github.com/ModOrganizer2/modorganizer/issues/938) but I know it's definitely not supposed to be cp1251, and if it's UTF-8, there's some mangling going on earlier in the sequence.