Add support for detailed error 126 diagnosis
https://github.com/ianpatt/f4se/commit/49c272156eb5d7ff05497784c73d7ee3632045b3
makes F4SE log the first module in a plugin's imports table that can't
be loaded when it hits error 126. This displays that information,
although it's not localised.

Also move a space inside an optional capture group so it works when it's
missing e.g. with the version of OBSE I was using in 2014.